backend/requirements.txt
# DocuMind AI — backend dependencies
# Phase 1: Core framework and database
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.0.0,<3.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9

# Phase 2: Document ingestion
PyMuPDF>=1.24.0,<2.0.0
pdfplumber>=0.11.0,<1.0.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
Pillow>=10.0.0,<11.0.0

# Phase 3: Processing (pre-installed for later phases)
presidio-analyzer>=2.2.0,<3.0.0
presidio-anonymizer>=2.2.0,<3.0.0
langdetect>=1.0.9,<2.0.0

# Phase 4: Retrieval
faiss-cpu>=1.8.0,<2.0.0
sentence-transformers>=3.0.0,<4.0.0
transformers>=4.40.0,<5.0.0
torch>=2.0.0
numpy>=1.26.0,<2.0.0
openai>=1.30.0,<2.0.0
